pyproject.toml
[project]
name = "veldra"
version = "0.1.0"
description = "Config-driven LightGBM analysis library scaffold"
readme = "README.md"
authors = [
    { name = "nbx", email = "nbx@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "joblib==1.5.3",
    "lightgbm==4.6.0",
    "numpy==2.4.2",
    "pandas==3.0.0",
    "pyarrow==23.0.0",
    "pydantic==2.12.5",
    "pyyaml==6.0.3",
    "scikit-learn==1.8.0",
]

[project.scripts]
veldra = "veldra:main"

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy==1.19.1",
    "pre-commit==4.5.1",
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "ruff==0.15.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:cacheprovider --basetemp=.pytest_tmp"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
